refactor(chat): extract StreamingManagerDelegate from ChatViewModel
Move the streaming session core out of ChatViewModel: the SSE collection job
and buffer, the throttled flush to UI state, the per-event dispatch
(handleStreamEvent + created/final milestones), app-foreground stream resume,
and network-error auto-reconnect. The send paths now build a request flow and
hand it to streamingManager.launchStream(); stop/pause/resume are thin
wrappers. No behavior change.
